+++ src/index.js
@@ -1,12 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import GlobalStyles from "./globalStyles";
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./components/App";
+import GlobalStyles from "./components/globalStyles";
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
     <GlobalStyles />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
