+++ src/utils/oneDecimal.js
@@ -1,0 +1,3 @@
+export default function oneDecimal(x) {
+  return Math.round(10 * x) / 10;
+}
